requirements.txt
pandas
xlsxwriter
jpype1
aspose-cells
openpyxl
numpy
sqlalchemy